+++ src/js/touchevents.tsx
@@ -2,6 +2,7 @@ import { addBreadcrumb, getCurrentHub } from '@sentry/core';
 import type { SeverityLevel } from '@sentry/types';
 import { logger } from '@sentry/utils';
 import * as React from 'react';
+import type { GestureResponderEvent} from 'react-native';
 import { StyleSheet, View } from 'react-native';
 
 import { createIntegration } from './integrations/factory';
@@ -60,6 +61,10 @@ interface ElementInstance {
   };
   memoizedProps?: Record<string, unknown>;
   return?: ElementInstance;
+}
+
+interface PrivateGestureResponderEvent extends GestureResponderEvent {
+  _targetInst?: ElementInstance;
 }
 
 /**
@@ -155,7 +160,7 @@ class TouchEventBoundary extends React.Component<TouchEventBoundaryProps> {
    * @param e
    */
   // eslint-disable-next-line complexity
-  private _onTouchStart(e: { _targetInst?: ElementInstance }): void {
+  private _onTouchStart(e: PrivateGestureResponderEvent): void {
     if (!e._targetInst) {
       return;
     }